feat: atributo pontos na classe nivel

+++ src/main/java/br/com/mavidsmile/mavidsmile/domains/Progresso.java
@@ -27,6 +27,6 @@ public class Progresso {
     private List<ProgressoPremio> premiosRecebidos;
 
     private int registros;
-    private double pontos;
+    private int pontos;
     private double porcentagemRegistros;
 }
